Once upon a time, Apple was so concerned about accuracy of floating point operations that they required all operations be performed in IEEE extended (80bit). This had the interesting side effect that a program which conformed to Apple's standards could not also conform to ANSI C (which required operations use double temporaries for double values). I find it somewhat ironic, therefore, that gcc on the mac doesn't even promote floats to doubles. FF needs the extra accuracy provided by a double (an extended type doesn't seem to provide additional assistance unfortunately). So on the mac we got extrema that weren't, which meant that remove overlap didn't.

+++ fontforge/bitmapdlg.c
@@ -482,7 +482,7 @@ static unichar_t *GenText(int32 *sizes,real scale) {
     pt = cret = galloc(i*10+1);
     for ( i=0; sizes[i]!=0; ++i ) {
 	if ( pt!=cret ) *pt++ = ',';
-	sprintf(pt,"%.1f",(sizes[i]&0xffff)*scale );
+	sprintf(pt,"%.1f",(double) ((sizes[i]&0xffff)*scale) );
 	pt += strlen(pt);
 	if ( pt[-1]=='0' && pt[-2]=='.' ) {
 	    pt -= 2;
